saas/requirements.txt
fastapi
uvicorn
openai
fastapi-clerk-auth
pydantic
pypdf
psycopg2-binary